consolidate fixtures
